feat: error boundaries on app and tab level #555
