+++ example-react-app/src/app/App.tsx
@@ -1,16 +1,15 @@
 import React from "react";
-import { Layout, Result } from "antd";
+import { Layout } from "antd";
 import { observer } from "mobx-react";
 import Login from "./login/Login";
 import Centered from "./common/Centered";
 import AppHeader from "./header/AppHeader";
 import { ContentArea } from "@haulmont/jmix-react-antd";
-import { useMainStore, Router, ErrorBoundary } from "@haulmont/jmix-react-core";
+import { useMainStore, Router } from "@haulmont/jmix-react-core";
 import CenteredLoader from "./CenteredLoader";
 import { AppMenu } from "./AppMenu";
 import "../routing";
 import styles from "./App.module.css";
-import { useIntl } from "react-intl";
 
 const routes = {
   "/": <ContentArea />,
@@ -27,52 +26,35 @@ const App = observer(() => {
 
   if (loginRequired) {
     return (
-      <AppErrorBoundary>
-        <Centered>
-          <Login />
-        </Centered>
-      </AppErrorBoundary>
+      <Centered>
+        <Login />
+      </Centered>
     );
   }
 
   return (
-    <AppErrorBoundary>
-      <Layout className={styles.mainLayout}>
-        <Layout.Header>
-          <AppHeader />
-        </Layout.Header>
-        <Layout className={styles.layoutContainer}>
-          <Layout.Sider
-            width={200}
-            breakpoint="sm"
-            collapsedWidth={0}
-            className={styles.layoutSider}
-          >
-            <AppMenu />
-          </Layout.Sider>
+    <Layout className={styles.mainLayout}>
+      <Layout.Header>
+        <AppHeader />
+      </Layout.Header>
+      <Layout className={styles.layoutContainer}>
+        <Layout.Sider
+          width={200}
+          breakpoint="sm"
+          collapsedWidth={0}
+          className={styles.layoutSider}
+        >
+          <AppMenu />
+        </Layout.Sider>
 
-          <Layout className={styles.layoutContent}>
-            <Layout.Content>
-              <Router global routes={routes} />
-            </Layout.Content>
-          </Layout>
+        <Layout className={styles.layoutContent}>
+          <Layout.Content>
+            <Router global routes={routes} />
+          </Layout.Content>
         </Layout>
       </Layout>
-    </AppErrorBoundary>
+    </Layout>
   );
 });
-
-const AppErrorBoundary = function(props) {
-  const intl = useIntl();
-
-  return (
-    <ErrorBoundary
-      message={intl.formatMessage({ id: "common.unknownAppError" })}
-      render={message => <Result status="warning" title={message} />}
-    >
-      {props.children}
-    </ErrorBoundary>
-  );
-};
 
 export default App;
