+++ example-react-app/src/index.tsx
@@ -10,7 +10,7 @@ import {
   JmixAppProvider,
   initializeApolloClient,
   Screens,
-  ScreensContext
+  ScreensContext, ErrorBoundary
 } from "@haulmont/jmix-react-core";
 import { I18nProvider, Modals } from "@haulmont/jmix-react-antd";
 import { initializeApp } from "@haulmont/jmix-rest";
@@ -34,6 +34,8 @@ import "dayjs/plugin/weekday";
 import "dayjs/plugin/localeData";
 import "dayjs/plugin/weekOfYear";
 import "dayjs/plugin/weekYear";
+import { Result } from "antd";
+import { useIntl } from "react-intl";
 
 initializeLocales();
 initializeTheme();
@@ -54,6 +56,19 @@ const client = initializeApolloClient({
 });
 
 const devScreens = new Screens();
+
+const AppErrorBoundary = function(props) {
+  const intl = useIntl();
+
+  return (
+    <ErrorBoundary
+      message={intl.formatMessage({ id: "common.unknownAppError" })}
+      render={message => <Result status="warning" title={message} />}
+    >
+      {props.children}
+    </ErrorBoundary>
+  );
+};
 
 ReactDOM.render(
   <JmixAppProvider
@@ -79,7 +94,9 @@ ReactDOM.render(
             }
             useInitialHook={useDevLogin}
           >
-            <App />
+            <AppErrorBoundary>
+              <App />
+            </AppErrorBoundary>
           </DevSupport>
         </IntlDocumentTitle>
       </I18nProvider>
@@ -87,4 +104,6 @@ ReactDOM.render(
   </JmixAppProvider>,
   document.getElementById("root") as HTMLElement
 );
+
+
 // registerServiceWorker();
